requirements.txt
anthropic>=0.40.0
openai>=1.0.0
ccxt>=4.4.0
replicate>=1.0.0
supabase>=2.10.0
python-dotenv>=1.0.0
pandas>=2.0.0
pandas-ta==0.4.71b0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
feedparser>=6.0.0
mplfinance>=0.12.10b0
redis>=5.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
